+++ index.html
@@ -250,7 +250,9 @@
         </li>
 
         <li>
-          <a href="https://wa.me/+543884215449"><i class="fa-brands fa-whatsapp"></i></a>
+          <a href="https://wa.me/+543884215449"
+            ><i class="fa-brands fa-whatsapp"></i
+          ></a>
         </li>
       </ul>
     </footer>
@@ -258,7 +260,7 @@
     <button class="dark-theme-btn" id="icon-theme">
       <i class="fa-solid fa-moon"></i>
     </button>
-    <script src="./js/index.js"></script>
+    <script type="module" src="./js/index.js"></script>
     <script src="./js/form.js"></script>
     <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
   </body>
